Install rpm-lockfile-prototype globally
The `rpm-lockfile-prototype` package was installed in the Python
directory of the user 'renovate'. As a result, running the 'renovate'
command as any other user will lead to a 'module not found' error for
'rpm_lockfile'. Adjusting the PYTHONPATH does not resolve this issue.
To work around this, we have installed `rpm-lockfile-prototype` globally.