Merge MintMaker and MPP-Renovate images
We have been maintaining two different images, one for our runners in
MPP, and another for Konflux. This commit merges the two images, which
can now be used in both environments and will reduce our maintenance
efforts (hopefully).

The most important changes with respect to the MPP image are:
- this image is based on ubi9-minimal (the MPP image was based on fedora)
- this image lacks python 3.8/3.10, as ubi9 does not have packages for
  these versions
- this image does not correction to rh-ca bundle mount

The moist important changes with respect to the previous version are:
- drop installation of `npm` and `nodejs` packages via the package manager
- manual installation of `nodejs`

There will be a follow-up change in the MPP deployment files. This
change must be merged first.

JIRA: [CWFHEALTH-3276]